Add files via upload
Problem Statement
Problem Statement 1:
You survey households in your area to find the average rent they are paying. Find the standard deviation from the following data:  1550,1550, 1700,  900,900, 850,  1000,1000, 950.

Problem Statement 2:
Find the variance for the following set of data representing trees in California (heights in feet): 3, 21, 98, 203, 17, 9

Problem Statement 3:
In a class on 100 students, 80 students passed in all subjects, 10 failed in one subject, 7 failed in two subjects and 3 failed in three subjects. Find the probability distribution of the variable for number of subjects a student from the given class has failed in.